+++ index.html
@@ -37,10 +37,10 @@
   </div>
 </template>
 
-<script src="js/animate.js"></script>
+<!-- <script src="js/animate.js"></script>
 <script src="js/time-format.js"></script>
 <script src="js/timer.js"></script>
-<script src="js/player.js"></script>
+<script src="js/player.js"></script> -->
 <script src="js/main.js"></script>
 
 </body>
