+++ js/data/data.js
@@ -1,0 +1,3 @@
+export {default as results} from './results';
+export {default as model} from './model';
+export {default as gameData} from './game';
