+++ js/modules/modules.js
@@ -1,0 +1,6 @@
+export {default as player} from './player';
+export {default as Application} from './application';
+export {default as game} from './game';
+export {default as Timer} from './timer-old';
+export {logo, mainTitle, title, timerMarkup} from './templates';
+export {renderUI, createElement} from './utils';
