+++ js/modules/timer.js
@@ -1,74 +1,62 @@
-const Timer = function (elem, duration, callback) {
-  this.elem = elem;
-  this.minutes = this.elem.querySelector('.timer-value-mins');
-  this.seconds = this.elem.querySelector('.timer-value-secs');
-  this.minVal = null;
-  this.secVall = null;
-  this.initDuration = duration;
-  this.duration = this.initDuration;
-  this.running = false;
-  this.callback = callback;
-  this.interval = null;
-  this.start = this.start.bind(this);
-  this._tick = this._tick.bind(this);
-  this._setValues = this._setValues.bind(this);
-  this.getLeftTime = this.getLeftTime.bind(this);
-  this.stop = this.stop.bind(this);
-  this.remove = this.remove.bind(this);
+import animate, {getAnimation} from './animate';
+import formatTime from './time-format';
+
+/*
+ * Окружность уменьшается за счет штриховки. Фактически, обводка состоит
+ * из одного длинного штриха, а пропуск за счет расстояния до следующего
+ * штриха. Задача правильной заливки состоит в том, чтобы правильно
+ * задать расстояние до следующего штриха.
+ *
+ * Из радиуса можно рассчитать длину окружности. При известной длине окружности,
+ * количестве шагов и номере текущего шага в анимации можно понять, на сколько
+ * нужно уменьшать длину окружности на текущем шаге. Для этого надо вычесть
+ * из нее длину одного шага умноженную на номер текущего шага.
+ *
+ * Длина окружности = 2πR
+ * Длина шага = Длина окружности / Количество шагов
+ * Пропуск = Длина шага * Номер шага
+ */
+const redrawCircle = (circle, radius, animation) => {
+  const length = 2 * Math.PI * radius;
+  const stepLength = length / animation.steps;
+  const lengthToClear = stepLength * animation.step;
+
+  circle.setAttributeNS(null, 'r', radius);
+  circle.setAttributeNS(null, 'stroke-dasharray', length.toString());
+  circle.setAttributeNS(null, 'stroke-dashoffset', lengthToClear.toString());
+
+  return circle;
 };
 
-Timer.prototype = {
-  start: function () {
-    if (this.running) {
-      return;
-    }
-    this._tick();
-    this.running = true;
-    this.interval = setInterval(this._tick, 1000);
-  },
-  _tick: function () {
-    this._setValues();
-    if (--this.duration < 0) {
-      this.stop();
-      if (typeof this.callback === 'function') {
-        this.callback();
-      }
-    }
-  },
-  _setValues: function () {
-    let minutes = parseInt(this.duration / 60, 10);
-    let seconds = parseInt(this.duration % 60, 10);
-    this.minVal = this.minutes.innerHTML = minutes < 10 ? '0' + minutes : minutes;
-    this.secVal = this.seconds.innerHTML = seconds < 10 ? '0' + seconds : seconds;
-  },
-  getLeftTime: function () {
-    return this.duration;
-  },
-  stop: function () {
-    if (!this.running) {
-      return;
-    }
-    clearInterval(this.interval);
-    this.running = false;
-  },
-  remove: function () {
-    this.elem = null;
-    this.minutes = null;
-    this.seconds = null;
-    this.minVal = null;
-    this.secVal = null;
-    this.initDuration = null;
-    this.duration = null;
-    this.running = null;
-    this.callback = null;
-    this.interval = null;
-    this.start = null;
-    this._tick = null;
-    this._setValues = null;
-    this.getLeftTime = null;
-    this.stop = null;
-    this.remove = null;
-  }
+
+const addLeadingZero = (val) => val < 10 ? `0${val}` : val;
+
+
+const redrawTimer = (timer, animation) => {
+  const total = animation.stepDuration * animation.steps;
+  const passed = animation.stepDuration * animation.step;
+  const timeLeft = formatTime(total, passed);
+
+  timer.querySelector('.timer-value-mins').textContent = addLeadingZero(timeLeft.minutes);
+  timer.querySelector('.timer-value-secs').textContent = addLeadingZero(timeLeft.seconds);
+
+  return timer;
 };
 
-export default Timer;
+
+export default (total = 120, onEnd) => {
+  const element = document.querySelector('.timer-line');
+  const radius = parseInt(element.getAttributeNS(null, 'r'), 10);
+  const timer = document.querySelector('.timer-value');
+
+  return animate(getAnimation(0, 1000, total), (animation) => {
+    redrawCircle(element, radius, animation);
+    redrawTimer(timer, animation);
+  }, () => {
+    if (typeof onEnd === 'function') {
+      onEnd();
+    }
+
+    timer.classList.add('timer-value--finished');
+  });
+};
