+++ js/modules/utils.js
@@ -1,10 +1,12 @@
-export const renderUI = (what) => {
+const renderUI = (what) => {
   let container = document.querySelector('.main');
   container.parentNode.replaceChild(what, container);
 };
 
-export const createElement = (markup) => {
+const createElement = (markup) => {
   let container = document.createElement('div');
   container.innerHTML = markup;
   return container.firstChild;
 };
+
+export {renderUI, createElement};
